SMART-8479: Remove e2fsprogs
We now compile and install the version of mke2fs from e2fsprogs
instead of the BusyBox one because BusyBox doesn't support ext3.
This is only done on the version of BusyBox used in the ug-uImage
for now.  The other chunks work okay, and we need to limit change.
If we don't do this, BusyBox will conflict with the other binary.